Move BGP code to new package
