+++ gnmi/bgp/gobgp.go
@@ -1,4 +1,4 @@
-package fakedevice
+package bgp
 
 import (
 	"context"
@@ -7,6 +7,7 @@ import (
 	"sync"
 
 	log "github.com/golang/glog"
+	"github.com/openconfig/lemming/gnmi/fakedevice"
 	"github.com/openconfig/lemming/gnmi/gnmiclient"
 	"github.com/openconfig/lemming/gnmi/oc"
 	"github.com/openconfig/lemming/gnmi/oc/ocpath"
@@ -60,7 +61,7 @@ func NewGoBGPTask() *reconciler.BuiltReconciler {
 //     ; however, if the global setting hasn't been set up, we actually need to erase the entirety of the applied config. This is because the watcher doesn't tell us this information.
 func startGoBGPFunc(ctx context.Context, yclient *ygnmi.Client) error {
 	b := &ocpath.Batch{}
-	bgpPath := ocpath.Root().NetworkInstance(DefaultNetworkInstance).Protocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP").Bgp()
+	bgpPath := ocpath.Root().NetworkInstance(fakedevice.DefaultNetworkInstance).Protocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP").Bgp()
 	b.AddPaths(
 		bgpPath.Global().As().Config().PathStruct(),
 		bgpPath.Global().RouterId().Config().PathStruct(),
@@ -75,7 +76,7 @@ func startGoBGPFunc(ctx context.Context, yclient *ygnmi.Client) error {
 
 	appliedRoot := &oc.Root{}
 	// appliedBgp is the SoT for BGP applied configuration. It is maintained locally by the task.
-	appliedBgp := appliedRoot.GetOrCreateNetworkInstance(DefaultNetworkInstance).GetOrCreateProtocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP").GetOrCreateBgp()
+	appliedBgp := appliedRoot.GetOrCreateNetworkInstance(fakedevice.DefaultNetworkInstance).GetOrCreateProtocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP").GetOrCreateBgp()
 	appliedBgp.PopulateDefaults()
 	var appliedBgpMu sync.Mutex
 
@@ -158,7 +159,7 @@ func startGoBGPFunc(ctx context.Context, yclient *ygnmi.Client) error {
 			if !ok {
 				return ygnmi.Continue
 			}
-			intended := rootVal.GetOrCreateNetworkInstance(DefaultNetworkInstance).GetOrCreateProtocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP").GetOrCreateBgp()
+			intended := rootVal.GetOrCreateNetworkInstance(fakedevice.DefaultNetworkInstance).GetOrCreateProtocol(oc.PolicyTypes_INSTALL_PROTOCOL_TYPE_BGP, "BGP").GetOrCreateBgp()
 
 			processBgp := func() {
 				log.V(1).Info("Processing BGP update")
